requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
ultralytics>=8.0.0
opencv-python>=4.6.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0
Pillow>=9.0.0
scikit-learn>=1.1.0

# Data processing and augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Dashboard and visualization
streamlit>=1.25.0
plotly>=5.10.0
gradio>=3.0.0

# XAI libraries
shap>=0.41.0
lime>=0.2.0.1
captum>=0.6.0
grad-cam>=1.4.0

# Text-to-speech and voice recognition
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
librosa>=0.9.0
soundfile>=0.10.0
gTTS>=2.5.1

# Model architectures
timm>=0.6.0
efficientnet-pytorch>=0.7.1

# Utilities
pyyaml>=6.0
tqdm>=4.64.0
pathlib2>=2.3.0
reportlab>=4.0.0
